improve viewer websocket connection robustness
This commit tries to fix a couple of viewer connection issues.
The viewer will now try to reconnect to the server every 5 seconds
after it got disconnected.
Furthermore in case the client has not received anything from the server
it will test its connection by sending a timestamp to the server which
will be echo'ed back.